updated URL
used 127.0.0.1 instead of 3001. It worked for the Grade Assignment function and sometimes for the Assign Assignment function. Still to check: should the backend data pop up in the dropdown, or does this have to coded in the frontend (app.tsx). In assign assignments dropdown it is coded in App.tsx. In the select student dropdown it isn't. The dropdown doesn't display any of the three names set in the Get request.